% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Lado servidor da aplicação}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Parâmetros internos para {shiny}.
NÃO REMOVA.}
}
\description{
Lado servidor da aplicação
}
