% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Interface do usuário da aplicação}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Parâmetro interno para `{shiny}`.
NÃO REMOVA.}
}
\description{
Interface do usuário da aplicação
}
