% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_analise_despesas.R
\name{m_analise_despesas_server}
\alias{m_analise_despesas_server}
\title{Servidor do módulo de análise de despesas}
\usage{
m_analise_despesas_server(id, dados_despesas)
}
\arguments{
\item{id}{Identificador único do módulo}

\item{dados_despesas}{Dados de despesas para análise}
}
\value{
Função do servidor do módulo
}
\description{
Servidor do módulo de análise de despesas
}
