% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_dashboard_principal.R
\name{m_dashboard_principal_server}
\alias{m_dashboard_principal_server}
\title{Servidor do módulo de dashboard principal}
\usage{
m_dashboard_principal_server(id)
}
\arguments{
\item{id}{Identificador único do módulo}
}
\value{
Função do servidor do módulo
}
\description{
Servidor do módulo de dashboard principal
}
