% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_despesas.R
\name{m_despesas_server}
\alias{m_despesas_server}
\title{Servidor do módulo de despesas}
\usage{
m_despesas_server(id, dados_despesas)
}
\arguments{
\item{id}{Identificador único do módulo}

\item{dados_despesas}{Dados de despesas para análise}
}
\value{
Função do servidor do módulo
}
\description{
Servidor do módulo de despesas
}
