% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_inicio.R
\name{m_inicio_server}
\alias{m_inicio_server}
\title{Servidor do módulo de início}
\usage{
m_inicio_server(id)
}
\arguments{
\item{id}{Identificador único do módulo}
}
\value{
Função do servidor do módulo
}
\description{
Servidor do módulo de início
}
