% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_receitas.R
\name{m_receitas_server}
\alias{m_receitas_server}
\title{Servidor do módulo de receitas}
\usage{
m_receitas_server(id, dados_receitas)
}
\arguments{
\item{id}{Identificador único do módulo}

\item{dados_receitas}{Dados de receitas para análise}
}
\value{
Função do servidor do módulo
}
\description{
Servidor do módulo de receitas
}
