% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_filtro_periodo.R
\name{sm_filtro_periodo_server}
\alias{sm_filtro_periodo_server}
\title{Servidor do submódulo de filtro de período}
\usage{
sm_filtro_periodo_server(id)
}
\arguments{
\item{id}{Identificador único do módulo}
}
\value{
Lista com elementos reativos para filtro de período
}
\description{
Servidor do submódulo de filtro de período
}
