% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_grafico_barras_empilhadas.R
\name{sm_grafico_barras_empilhadas_server}
\alias{sm_grafico_barras_empilhadas_server}
\title{Servidor do submódulo de gráficos de barras empilhadas}
\usage{
sm_grafico_barras_empilhadas_server(
  id,
  dados,
  filtro_periodo,
  data_inicial,
  data_final,
  choices,
  max_unicos_i = 20,
  total = "total.pago",
  data = "data.doc.pagto",
  comecoTitulo = "Análise"
)
}
\arguments{
\item{id}{Identificador único do módulo}

\item{dados}{Dados para análise}

\item{filtro_periodo}{Reativo com tipo de filtro de período}

\item{data_inicial}{Reativo com data inicial (para filtro personalizado)}

\item{data_final}{Reativo com data final (para filtro personalizado)}

\item{choices}{Lista de opções para agrupamento}

\item{max_unicos_i}{Máximo de categorias únicas antes de agrupar em "Outros"}

\item{total}{Nome da coluna com valores totais}

\item{data}{Nome da coluna com datas}

\item{comecoTitulo}{Início do título do gráfico}
}
\value{
Função do servidor do módulo
}
\description{
Servidor do submódulo de gráficos de barras empilhadas
}
