% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_grafico_barras_empilhadas.R
\name{sm_grafico_barras_empilhadas_server}
\alias{sm_grafico_barras_empilhadas_server}
\title{Servidor do submódulo de gráficos de barras empilhadas}
\usage{
sm_grafico_barras_empilhadas_server(
  id,
  dados,
  choices,
  total = "total.pago",
  data = "data.doc.pagto",
  comecoTitulo = "Análise"
)
}
\arguments{
\item{id}{Identificador único do módulo}

\item{dados}{Dados para análise}

\item{choices}{Lista de opções para agrupamento}

\item{total}{Nome da coluna com valores totais}

\item{data}{Nome da coluna com datas}

\item{comecoTitulo}{Início do título do gráfico}
}
\value{
Função do servidor do módulo
}
\description{
Servidor do submódulo de gráficos de barras empilhadas
}
